Refactor MarkdownWriters
* Put each vital definition in its own trait
* Fix some obvious shortcomings
