+++ hugo/src/main/scala/com/ossuminc/riddl/hugo/writers/SagaWriter.scala
@@ -1,0 +1,29 @@
+package com.ossuminc.riddl.hugo.writers
+
+import com.ossuminc.riddl.hugo.writers.MarkdownWriter
+import com.ossuminc.riddl.language.AST.*
+
+trait SagaWriter { this: MarkdownWriter =>
+
+  private def emitSagaSteps(actions: Seq[SagaStep]): this.type = {
+    h2("Saga Actions")
+    actions.foreach { step =>
+      h3(step.identify)
+      emitShortDefDoc(step)
+      list(typeOfThing = "Do Statements", step.doStatements.map(_.format), 4)
+      list(typeOfThing = "Undo Statements", step.doStatements.map(_.format), 4)
+    }
+    this
+  }
+
+  def emitSaga(saga: Saga, parents: Seq[String]): Unit = {
+    containerHead(saga, "Saga")
+    emitDefDoc(saga, parents)
+    emitOptions(saga.options)
+    emitInputOutput(saga.input, saga.output)
+    emitSagaSteps(saga.sagaSteps)
+    emitUsage(saga)
+    emitTerms(saga.terms)
+    emitIndex("Saga", saga, parents)
+  }
+}
